further prep of the contrib + tests code for monolithic build mode Win32/Win64 builds.

+++ zlibWrapper/examples/example_original.c
@@ -452,7 +452,7 @@ static void test_large_inflate(unsigned char *compr, size_t comprLen, unsigned c
     CHECK_ERR(err, "inflateEnd");
 
     if (d_stream.total_out != 2*uncomprLen + diff) {
-        fprintf(stderr, "bad large inflate: %" PRIu64 "\n", (uint64_t)d_stream.total_out);
+        fprintf(stderr, "bad large inflate: %zu\n", d_stream.total_out);
         exit(1);
     } else {
         printf("large_inflate(): OK\n");
